requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
django-mysql==4.7.1
djangorestframework==3.14.0
mysqlclient==2.1.1
pika==1.3.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
